Create api for adding products to cart
